Release 2.3.0-beta1

New features:
- Added schema manager including migrations.
- Search index population has been rewritten.
- Insert tag handling has been rewritten.
- Sub-headline in input screens can be manually defined.
- It is now possible to add an URL-fragment to the sorting
  links list elements and modules.
- Sorting links in the list templates can now be generated
  via methods from within the template.
- Support for native routing (Since Contao 4.10) is now
  officially supported - including priority handling.
- Session handling has been moved from Contao Session to
  Symfony session - this reduces the amount of needed
  Cookies.
- Invariant attributes are not hidden for variant items
  anymore but shown as read only instead.
- Attribute settings in input screens now allow to select
  a widget template for backend and frontend-editing.
- Listing in Backend can now be grouped by calender week.
- Translation handling has been converted from legacy
  language arrays (as known from TYPOlight days) and
  contao-community-alliance/translator to usage of the
  symfony/translator where possible.
- The backend routes have been changed to use custom
  controllers. This removes the requirement to define
  access rights in multiple places. Now only the rights
  in MetaModels have to be defined.

Fixes:
- Locale handling has been adjusted to Contao.
- Child MetaModels can now also handle variants.
- Many micro-fixes

Various:
- The check pipeline has been switched from phpcq 1.3 to
  phpcq 2.0.
- The whole codebase has been improved to fix found issues.

+++ src/Attribute/BaseSimple.php
@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2022 The MetaModels team.
+ * (c) 2012-2024 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -18,7 +18,7 @@
  * @author     Sven Baumann <baumann.sv@gmail.com>
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Marc Reimann <reimann@mediendepot-ruhr.de>
- * @copyright  2012-2022 The MetaModels team.
+ * @copyright  2012-2024 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -26,9 +26,16 @@
 namespace MetaModels\Attribute;
 
 use Contao\System;
+use Doctrine\DBAL\ArrayParameterType;
 use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\Exception;
 use MetaModels\Helper\TableManipulator;
 use MetaModels\IMetaModel;
+
+use function array_map;
+use function is_array;
+use function serialize;
+use function str_replace;
 
 /**
  * Reference implementation for Simple attributes.
@@ -39,6 +46,8 @@ use MetaModels\IMetaModel;
  */
 class BaseSimple extends Base implements ISimple
 {
+    use ManagedAttributeTrait;
+
     /**
      * Database connection.
      *
@@ -58,13 +67,13 @@ class BaseSimple extends Base implements ISimple
      *
      * Note that you should not use this directly but use the factory classes to instantiate attributes.
      *
-     * @param IMetaModel       $objMetaModel     The MetaModel instance this attribute belongs to.
-     * @param array            $arrData          The information array, for attribute information, refer to
-     *                                           documentation of table tl_metamodel_attribute and documentation of the
-     *                                           certain attribute classes for information what values are understood.
-     * @param Connection       $connection       The database connection.
-     *
-     * @param TableManipulator $tableManipulator Table manipulator instance.
+     * @param IMetaModel            $objMetaModel     The MetaModel instance this attribute belongs to.
+     * @param array                 $arrData          The information array, for attribute information, refer to
+     *                                                documentation of table tl_metamodel_attribute and documentation
+     *                                                of the certain attribute classes for information what values are
+     *                                                understood.
+     * @param Connection|null       $connection       The database connection.
+     * @param TableManipulator|null $tableManipulator Table manipulator instance.
      */
     public function __construct(
         IMetaModel $objMetaModel,
@@ -82,7 +91,9 @@ class BaseSimple extends Base implements ISimple
             );
             // @codingStandardsIgnoreEnd
             $connection = System::getContainer()->get('database_connection');
+            assert($connection instanceof Connection);
         }
+        $this->connection = $connection;
 
         if (null === $tableManipulator) {
             // @codingStandardsIgnoreStart
@@ -93,9 +104,8 @@ class BaseSimple extends Base implements ISimple
             // @codingStandardsIgnoreEnd
 
             $tableManipulator = System::getContainer()->get('metamodels.table_manipulator');
+            assert($tableManipulator instanceof TableManipulator);
         }
-
-        $this->connection       = $connection;
         $this->tableManipulator = $tableManipulator;
     }
 
@@ -104,26 +114,38 @@ class BaseSimple extends Base implements ISimple
      *
      * This tells the attribute to perform any actions that must be done to correctly initialize the new value
      * and to perform any action to undo the changes that had been done for the previous value.
-     * i.e.: when an attribute type needs columns in an an auxiliary table, these will have to be updated herein.
+     * i.e.: when an attribute type needs columns in an auxiliary table, these will have to be updated herein.
      *
      * This method may throw an exception, when the new value is invalid or any problems appear, the MetaModelAttribute
      * will then keep the old meta value.
      *
      * @param string $strMetaName Name of the meta information that shall be updated.
+     * @param mixed $varNewValue The new value for this meta information.
      *
-     * @param mixed  $varNewValue The new value for this meta information.
+     * @return IAttribute The instance of this attribute, to support chaining.
      *
-     * @return \MetaModels\Attribute\IAttribute The instance of this attribute, to support chaining.
+     * @throws Exception
+     *
+     * @deprecated Implement schema generators instead.
      */
     public function handleMetaChange($strMetaName, $varNewValue)
     {
-        // By default we accept any change of meta information.
-        if ($strMetaName == 'colname') {
-            if ($this->get($strMetaName) != $varNewValue) {
-                $this->renameColumn($varNewValue);
-            }
+        if ($this->isManagedAttribute($this->get('type'))) {
+            $this->triggerDeprecationShouldNotCallManaged(static::class, __METHOD__);
             return $this;
         }
+
+        // By default, we accept any change of meta information.
+        if ($strMetaName === 'colname') {
+            if ($this->get($strMetaName) !== $varNewValue) {
+                /** @psalm-suppress DeprecatedMethod */
+                $this->renameColumn($varNewValue);
+            }
+
+            return $this;
+        }
+
+        /** @psalm-suppress DeprecatedMethod */
         return parent::handleMetaChange($strMetaName, $varNewValue);
     }
 
@@ -133,6 +155,8 @@ class BaseSimple extends Base implements ISimple
      * @param mixed $arrValues The values to be stored into database. Mapping is item id=>value.
      *
      * @return void
+     *
+     * @throws Exception
      */
     public function setDataFor($arrValues)
     {
@@ -146,12 +170,14 @@ class BaseSimple extends Base implements ISimple
                 ->set('t.' . $strColName, ':' . $strColName)
                 ->setParameter($strColName, is_array($varData) ? serialize($varData) : $varData)
                 ->setParameter('id', $intId)
-                ->execute();
+                ->executeQuery();
         }
     }
 
     /**
      * {@inheritDoc}
+     *
+     * @throws Exception
      */
     public function getFilterOptions($idList, $usedOnly, &$arrCount = null)
     {
@@ -168,25 +194,26 @@ class BaseSimple extends Base implements ISimple
                 ->where('t.id IN (:ids)')
                 ->groupBy('t.' . $strCol)
                 ->orderBy('MIN(FIELD(t.id, :ids))')
-                ->setParameter('ids', $idList, Connection::PARAM_STR_ARRAY)
-                ->execute();
+                ->setParameter('ids', $idList, ArrayParameterType::STRING)
+                ->executeQuery();
         } else {
             $statement = $this->connection->createQueryBuilder()
                 ->select('t.' . $strCol . ', COUNT(t.' . $strCol . ') as mm_count')
                 ->from($this->getMetaModel()->getTableName(), 't')
                 ->groupBy('t.' . $strCol)
                 ->orderBy('t.' . $strCol)
-                ->execute();
+                ->executeQuery();
         }
 
         $arrResult = [];
-        while ($objRow = $statement->fetch(\PDO::FETCH_OBJ)) {
+        while ($objRow = $statement->fetchAssociative()) {
             if (is_array($arrCount)) {
-                $arrCount[$objRow->$strCol] = $objRow->mm_count;
+                $arrCount[$objRow[$strCol]] = $objRow['mm_count'];
             }
 
-            $arrResult[$objRow->$strCol] = $objRow->$strCol;
+            $arrResult[$objRow[$strCol]] = $objRow[$strCol];
         }
+
         return $arrResult;
     }
 
@@ -194,20 +221,23 @@ class BaseSimple extends Base implements ISimple
      * {@inheritdoc}
      *
      * This base implementation does a plain SQL sort by native value as defined by MySQL.
+     *
+     * @throws Exception
      */
     public function sortIds($idList, $strDirection)
     {
         // Base implementation, do a simple sorting on given column.
-        $idList = $this->connection->createQueryBuilder()
+        $statement = $this->connection
+            ->createQueryBuilder()
             ->select('t.id')
             ->from($this->getMetaModel()->getTableName(), 't')
             ->where('t.id IN (:ids)')
-            ->setParameter('ids', $idList, Connection::PARAM_STR_ARRAY)
+            ->setParameter('ids', $idList, ArrayParameterType::STRING)
             ->orderBy('t.' . $this->getColName(), $strDirection)
-            ->execute()
-            ->fetchAll(\PDO::FETCH_COLUMN, 'id');
+            ->executeQuery();
 
-        return $idList;
+        // Return value list as list<mixed>, parent function wants a list<string> so we make a cast.
+        return array_map(static fn (mixed $value) => (string) $value, $statement->fetchFirstColumn());
     }
 
     /**
@@ -218,19 +248,25 @@ class BaseSimple extends Base implements ISimple
      *
      * @param string $strPattern The text to search for. This may contain wildcards.
      *
-     * @return int[] the ids of matching items.
+     * @return list<string> The ids of matching items.
+     *
+     * @throws Exception
      */
     public function searchFor($strPattern)
     {
         // Base implementation, do a simple search on given column.
-        $strPattern = str_replace(array('*', '?'), array('%', '_'), $strPattern);
-        return $this->connection->createQueryBuilder()
+        $strPattern = str_replace(['*', '?'], ['%', '_'], $strPattern);
+
+        $statement = $this->connection
+            ->createQueryBuilder()
             ->select('t.id')
             ->from($this->getMetaModel()->getTableName(), 't')
             ->where('t.' . $this->getColName() . ' LIKE :pattern')
             ->setParameter('pattern', $strPattern)
-            ->execute()
-            ->fetchAll(\PDO::FETCH_COLUMN, 'id');
+            ->executeQuery();
+
+        // Return value list as list<mixed>, parent function wants a list<string> so we make a cast.
+        return array_map(static fn (mixed $value) => (string) $value, $statement->fetchFirstColumn());
     }
 
     /**
@@ -240,9 +276,15 @@ class BaseSimple extends Base implements ISimple
      * Deriving classes SHOULD override this function.
      *
      * @return string 'blob NULL'
+     *
+     * @deprecated Implement schema generators instead - see #1267.
      */
     public function getSQLDataType()
     {
+        if ($this->isManagedAttribute($this->get('type'))) {
+            $this->triggerDeprecationShouldNotCallManaged(static::class, __METHOD__);
+        }
+
         return 'blob NULL';
     }
 
@@ -250,10 +292,23 @@ class BaseSimple extends Base implements ISimple
      * Create auxiliary data like a column in the MetaModel table or references in another table etc.
      *
      * @return void
+     *
+     * @deprecated Implement schema generators instead - see #1267.
      */
     public function destroyAUX()
     {
+        if ($this->isManagedAttribute($this->get('type'))) {
+            $this->triggerDeprecationShouldNotCallManaged(static::class, __METHOD__);
+
+            return;
+        }
+
+        $this->triggerDeprecationIsUnmanagedAttribute(static::class, __METHOD__);
+
+        /** @psalm-suppress DeprecatedMethod */
         parent::destroyAUX();
+
+        /** @psalm-suppress DeprecatedMethod */
         $this->deleteColumn();
     }
 
@@ -261,10 +316,25 @@ class BaseSimple extends Base implements ISimple
      * Delete all auxiliary data like a column in the MetaModel table or references in another table etc.
      *
      * @return void
+     *
+     * @throws Exception
+     *
+     * @deprecated Implement schema generators instead - see #1267.
      */
     public function initializeAUX()
     {
+        if ($this->isManagedAttribute($this->get('type'))) {
+            $this->triggerDeprecationShouldNotCallManaged(static::class, __METHOD__);
+
+            return;
+        }
+
+        $this->triggerDeprecationIsUnmanagedAttribute(static::class, __METHOD__);
+
+        /** @psalm-suppress DeprecatedMethod */
         parent::initializeAUX();
+
+        /** @psalm-suppress DeprecatedMethod */
         $this->createColumn();
     }
 
@@ -274,10 +344,23 @@ class BaseSimple extends Base implements ISimple
      * You have to override this function in field types, when you want to have multi column structure etc.
      *
      * @return void
+     *
+     * @throws Exception
+     *
+     * @deprecated Implement schema generators instead - see #1267.
      */
     public function createColumn()
     {
+        if ($this->isManagedAttribute($this->get('type'))) {
+            $this->triggerDeprecationShouldNotCallManaged(static::class, __METHOD__);
+
+            return;
+        }
+
+        $this->triggerDeprecationIsUnmanagedAttribute(static::class, __METHOD__);
+
         if ($this->getColName()) {
+            /** @psalm-suppress DeprecatedMethod */
             $this->tableManipulator->createColumn(
                 $this->getMetaModel()->getTableName(),
                 $this->getColName(),
@@ -290,10 +373,22 @@ class BaseSimple extends Base implements ISimple
      * Removes the underlying database structure for this field.
      *
      * @return void
+     *
+     * @throws Exception
+     *
+     * @deprecated Implement schema generators instead - see #1267.
      */
     public function deleteColumn()
     {
-        $schemaManager = $this->connection->getSchemaManager();
+        if ($this->isManagedAttribute($this->get('type'))) {
+            $this->triggerDeprecationShouldNotCallManaged(static::class, __METHOD__);
+
+            return;
+        }
+
+        $this->triggerDeprecationIsUnmanagedAttribute(static::class, __METHOD__);
+
+        $schemaManager = $this->connection->createSchemaManager();
         $columns       = $schemaManager->listTableColumns($this->getMetaModel()->getTableName());
 
         // Try to delete the column. If it does not exist as we can assume it has been deleted already then.
@@ -308,15 +403,28 @@ class BaseSimple extends Base implements ISimple
      * @param string $strNewColumnName The new column name.
      *
      * @return void
+     *
+     * @throws Exception
+     *
+     * @deprecated Implement schema generators instead - see #1267.
      */
     public function renameColumn($strNewColumnName)
     {
+        if ($this->isManagedAttribute($this->get('type'))) {
+            $this->triggerDeprecationShouldNotCallManaged(static::class, __METHOD__);
+
+            return;
+        }
+
+        $this->triggerDeprecationIsUnmanagedAttribute(static::class, __METHOD__);
+
         $this->tableManipulator->checkColumnName($strNewColumnName);
 
-        $schemaManager = $this->connection->getSchemaManager();
+        $schemaManager = $this->connection->createSchemaManager();
         $columns       = $schemaManager->listTableIndexes($this->getMetaModel()->getTableName());
 
         if ($this->getColName() && isset($columns[$this->getColName()])) {
+            /** @psalm-suppress DeprecatedMethod */
             $this->tableManipulator->renameColumn(
                 $this->getMetaModel()->getTableName(),
                 $this->getColName(),
@@ -326,6 +434,7 @@ class BaseSimple extends Base implements ISimple
         } else {
             $strBackupColName = $this->getColName();
             $this->set('colname', $strNewColumnName);
+            /** @psalm-suppress DeprecatedMethod */
             $this->createColumn();
             $this->set('colname', $strBackupColName);
         }
@@ -348,10 +457,14 @@ class BaseSimple extends Base implements ISimple
      *
      * @param mixed $value The input value.
      *
-     * @return string
+     * @return string|null
      */
     public function serializeData($value)
     {
-        return $value;
+        if (empty($value)) {
+            return null;
+        }
+
+        return (string) $value;
     }
 }
